// SDKs/storecove.com/2.0.1/typescript/src/sdk/models/shared/invoiceline.ts
export class InvoiceLine extends SpeakeasyBase {
  @SpeakeasyMetadata({ data: "json, name=accountingCost" })
  accountingCost?: string;

  @SpeakeasyMetadata({ data: "json, name=additionalItemProperties", elemType: AdditionalItemProperty })
  additionalItemProperties?: AdditionalItemProperty[];

  @SpeakeasyMetadata({ data: "json, name=allowanceCharge" })
  allowanceCharge?: number;

  @SpeakeasyMetadata({ data: "json, name=amountExcludingVat" })
  amountExcludingVat: number;

  @SpeakeasyMetadata({ data: "json, name=buyersItemIdentification" })
  buyersItemIdentification?: string;

  @SpeakeasyMetadata({ data: "json, name=delivery" })
  delivery?: Delivery;

  @SpeakeasyMetadata({ data: "json, name=description" })
  description?: string;

  @SpeakeasyMetadata({ data: "json, name=invoicePeriod" })
  invoicePeriod?: string;

  @SpeakeasyMetadata({ data: "json, name=itemPrice" })
  itemPrice?: number;

  @SpeakeasyMetadata({ data: "json, name=lineId" })
  lineId?: string;

  @SpeakeasyMetadata({ data: "json, name=name" })
  name?: string;

  @SpeakeasyMetadata({ data: "json, name=orderLineReferenceLineId" })
  orderLineReferenceLineId?: string;

  @SpeakeasyMetadata({ data: "json, name=quantity" })
  quantity?: number;

  @SpeakeasyMetadata({ data: "json, name=quantityUnitCode" })
  quantityUnitCode?: InvoiceLineQuantityUnitCodeEnum;

  @SpeakeasyMetadata({ data: "json, name=sellersItemIdentification" })
  sellersItemIdentification?: string;

  @SpeakeasyMetadata({ data: "json, name=standardItemIdentification" })
  standardItemIdentification?: string;

  @SpeakeasyMetadata({ data: "json, name=standardItemIdentificationSchemeAgencyId" })
  standardItemIdentificationSchemeAgencyId?: string;

  @SpeakeasyMetadata({ data: "json, name=standardItemIdentificationSchemeId" })
  standardItemIdentificationSchemeId?: string;

  @SpeakeasyMetadata({ data: "json, name=tax" })
  tax?: Tax;

  @SpeakeasyMetadata({ data: "json, name=taxes_duties_fees", elemType: Tax })
  taxesDutiesFees?: Tax[];
}